booking controller

+++ src/controller/appointment.controller/appointment.controller.js
@@ -3,17 +3,24 @@ var { StatusCodes } = require('http-status-codes');
 // appointment add controller 
 exports.appointmentAdd = async (req, res) => {
     try {
+        if (!req.userInfo.isAdmin) {
+            return res.status(StatusCodes.UNAUTHORIZED).send({
+                success: false,
+                message: "Unauthorized Access"
+            })
+        }
         const { appointment_title, appointment_description, appointment_status, appointment_data, appointment_time } = req.body;
         var _appointment = await new appointmentModel({
-            appointment_data, appointment_time, appointment_status, appointment_title, appointment_description
+            user: req.userInfo._id, appointment_data, appointment_time, appointment_status, appointment_title, appointment_description
         })
         await _appointment.save();
-        res.status(StatusCodes.OK).send({
+        return res.status(StatusCodes.OK).send({
             success: true,
             message: "Appointment has been added"
         })
     } catch (error) {
-        res.status(StatusCodes.INTERNAL_SERVER_ERROR).send({
+        console.log(error)
+        return res.status(StatusCodes.INTERNAL_SERVER_ERROR).send({
             success: false,
             message: "Intername server error !!"
         })
@@ -23,13 +30,19 @@ exports.appointmentAdd = async (req, res) => {
 exports.appointmentUpdate = async (req, res) => {
     try {
         const { appointment_title, appointment_description, appointment_status, appointment_data, appointment_time } = req.body;
+        if (!req.userInfo.isAdmin) {
+            return res.status(StatusCodes.UNAUTHORIZED).send({
+                success: false,
+                message: "Unauthorized Access"
+            })
+        }
         var _appointmentUpdate = await appointmentModel.findOneAndUpdate({ _id: req.params.id }, req?.body)
-        res.status(StatusCodes.OK).send({
+        return res.status(StatusCodes.OK).send({
             success: true,
             message: "Appointment has been updated"
         })
     } catch (error) {
-        res.status(StatusCodes.INTERNAL_SERVER_ERROR).send({
+        return res.status(StatusCodes.INTERNAL_SERVER_ERROR).send({
             success: false,
             message: "Intername server error !!"
         })
@@ -39,13 +52,19 @@ exports.appointmentUpdate = async (req, res) => {
 //apportment delete controller 
 exports.appointmentDelete = async (req, res) => {
     try {
+        if (!req.userInfo.isAdmin) {
+            return res.status(StatusCodes.UNAUTHORIZED).send({
+                success: false,
+                message: "Unauthorized Access"
+            })
+        }
         var _appointmentDelete = await appointmentModel.findOneAndDelete({ _id: req.params.id })
-        res.status(StatusCodes.OK).send({
+        return res.status(StatusCodes.OK).send({
             success: true,
             message: "Appointment has been deleted"
         })
     } catch (error) {
-        res.status(StatusCodes.INTERNAL_SERVER_ERROR).send({
+        return res.status(StatusCodes.INTERNAL_SERVER_ERROR).send({
             success: false,
             message: "Intername server error !!"
         })
@@ -56,12 +75,12 @@ exports.appointmentDelete = async (req, res) => {
 exports.appointmentGet = async (req, res) => {
     try {
         var _appointmentGEt = await appointmentModel.find();
-        res.status(StatusCodes.OK).send({
+        return res.status(StatusCodes.OK).send({
             success: true,
-            data:_appointmentGEt
+            data: _appointmentGEt
         })
     } catch (error) {
-        res.status(StatusCodes.INTERNAL_SERVER_ERROR).send({
+        return res.status(StatusCodes.INTERNAL_SERVER_ERROR).send({
             success: false,
             message: "Intername server error !!"
         })
